moved inline diff files into diff
