+++ src/vs/workbench/contrib/notebook/browser/diff/inlineDiff/notebookInlineDiffWidget.ts
@@ -3,21 +3,21 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import * as DOM from '../../../../../base/browser/dom.js';
-import { CodeWindow } from '../../../../../base/browser/window.js';
-import { Event } from '../../../../../base/common/event.js';
-import { Disposable } from '../../../../../base/common/lifecycle.js';
-import { EditorExtensionsRegistry } from '../../../../../editor/browser/editorExtensions.js';
-import { MenuId } from '../../../../../platform/actions/common/actions.js';
-import { IInstantiationService } from '../../../../../platform/instantiation/common/instantiation.js';
-import { NotebookTextModel } from '../../common/model/notebookTextModel.js';
-import { NotebookDiffEditorInput } from '../../common/notebookDiffEditorInput.js';
-import { NotebookInlineDiffDecorationContribution } from '../contrib/inlineDiff/notebookInlineDiff.js';
-import { INotebookEditorOptions } from '../notebookBrowser.js';
-import { NotebookEditorExtensionsRegistry } from '../notebookEditorExtensions.js';
-import { NotebookEditorWidget } from '../notebookEditorWidget.js';
-import { NotebookOptions } from '../notebookOptions.js';
-import { IBorrowValue, INotebookEditorService } from '../services/notebookEditorService.js';
+import * as DOM from '../../../../../../base/browser/dom.js';
+import { CodeWindow } from '../../../../../../base/browser/window.js';
+import { Event } from '../../../../../../base/common/event.js';
+import { Disposable } from '../../../../../../base/common/lifecycle.js';
+import { EditorExtensionsRegistry } from '../../../../../../editor/browser/editorExtensions.js';
+import { MenuId } from '../../../../../../platform/actions/common/actions.js';
+import { IInstantiationService } from '../../../../../../platform/instantiation/common/instantiation.js';
+import { NotebookTextModel } from '../../../common/model/notebookTextModel.js';
+import { NotebookDiffEditorInput } from '../../../common/notebookDiffEditorInput.js';
+import { NotebookInlineDiffDecorationContribution } from './notebookInlineDiff.js';
+import { INotebookEditorOptions } from '../../notebookBrowser.js';
+import { NotebookEditorExtensionsRegistry } from '../../notebookEditorExtensions.js';
+import { NotebookEditorWidget } from '../../notebookEditorWidget.js';
+import { NotebookOptions } from '../../notebookOptions.js';
+import { IBorrowValue, INotebookEditorService } from '../../services/notebookEditorService.js';
 
 export class NotebookInlineDiffWidget extends Disposable {
