+++ src/vs/workbench/contrib/notebook/common/notebookCommon.ts
@@ -1034,7 +1034,7 @@ export const NotebookSetting = {
 	cellFailureDiagnostics: 'notebook.cellFailureDiagnostics',
 	outputBackupSizeLimit: 'notebook.backup.sizeLimit',
 	multiCursor: 'notebook.multiCursor.enabled',
-	markupFontFamily: 'notebook.markup.fontFamily'
+	markupFontFamily: 'notebook.markup.fontFamily',
 } as const;
 
 export const enum CellStatusbarAlignment {
